raw-win32 driver reopen support
This is the support for reopen, for win32.  Note that there is a key
difference in the win32 reopen, namely that it is not guaranteed safe
like all the other drivers.  Attempts are made to reopen the file, or
open the file in a new handle prior to closing the old handle.  However,
this is not always feasible, and there are times when we must close the
existing file and then open the new file, breaking the transactional
nature of the reopen.

Original author: Jeff Cody
http://lists.gnu.org/archive/html/qemu-devel/2012-09/msg02007.html

How to test the reopen feature works:

$qemuPath = ".\qemu-img.exe"
$format = "qcow2"
$qcowImagePath = ".\test1.${format}"
$qcowImageSnapPath = "${qcowImagePath}.snap"
& $qemuPath --version
& $qemuPath create -f $format $qcowImagePath 1G
& $qemuPath create -f $format -b $qcowImagePath $qcowImageSnapPath
& $qemuPath commit $qcowImageSnapPath

qemu-img command should succeed with message: Image commited.
If the command fails, the patch does not work. The error message
would be:

.\qemu-img.exe: Block format 'file' used by node '#block238' does no
support reopening files